// src/app/models/user.ts
import {CompanyModel} from "./company.model";

export class User{
  lastName: string;
  userStatus: string;
  jobTitle: string;
  statusUpdatedAt: string;
  mobile: string;
  emailNotificationsEnabled: boolean;
  description: string;
  telephone: string;
  enabled: boolean;
  firstName: string;
  skypeId: string;
  avatarId: string;
  location: string;
  company: CompanyModel;
  id: string;
  email: string;
}